refactor: Seq Assign/ResetValue as ulong (R-C4)
Align sequential reset immediates with the 64-bit value world.

+++ src/SharpHdl.Core/Model/Signal.cs
@@ -43,7 +43,7 @@ public class Signal : Expr
 		}
 	}
 
-	public void Assign(uint resetValue, Expr next)
+	public void Assign(ulong resetValue, Expr next)
 	{
 		if (next.GetWidth() != Width)
 		{
